samples: hello_ei: Refactor CMake build system with dual compilation modes
Restructure Edge Impulse library integration to support two build
configurations:

1. Compile into app (EDGE_IMPULSE_COMPILE_IN_TREE=y):
   - Uses FetchContent to compile EI sources directly into application
   - Faster incremental builds and better optimization (LTO)
   - Warning suppression applied per-file to EI sources only

2. Standalone library (EDGE_IMPULSE_COMPILE_IN_TREE=n):
   - Uses ExternalProject to build EI as separate static library
   - Better build caching and reusability across multiple apps
   - Isolated build context with independent compiler flags

Changes:
- Split monolithic CMakeLists.txt into modular files:
  * ei_sdk_utils.cmake: URI processing and API key handling
  * ei_compile_in_app.cmake: FetchContent-based integration
  * ei_standalone_lib.cmake: ExternalProject-based integration
- Add Kconfig option EDGE_IMPULSE_COMPILE_IN_TREE
- Simplify main CMakeLists.txt to conditional include logic
- Update prj.conf and main.c with minor improvements
- Reduce number of include directories

+++ samples/edge_impulse/hello_ei/src/main.c
@@ -38,11 +38,11 @@ static void print_inference_result(const ei_impulse_result_t *result, int64_t du
 
 	for (size_t i = 0; i < EI_CLASSIFIER_LABEL_COUNT; i++) {
 		LOG_INF("%s => %.5f", result->classification[i].label,
-		        result->classification[i].value);
+			(double)result->classification[i].value);
 	}
 
 #if EI_CLASSIFIER_HAS_ANOMALY
-	LOG_INF("anomaly: %.5f", result->anomaly);
+	LOG_INF("anomaly: %.5f", (double)result->anomaly);
 #endif
 
 	LOG_INF("=== Inference time profiling ===");
